services/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
prisma==0.11.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
sqlalchemy
psycopg2-binary
requests
jinja2
python-jose
